Fix sm300d2 preamble sensing and checksum calculation v3

+++ esphome/components/sm300d2/sm300d2.h
@@ -23,7 +23,7 @@ class SM300D2Sensor : public PollingComponent, public uart::UARTDevice {
   void dump_config() override;
 
  protected:
-  uint16_t sm300d2_checksum_(uint8_t *ptr);
+  uint8_t sm300d2_checksum_(uint8_t *ptr);
 
   sensor::Sensor *co2_sensor_{nullptr};
   sensor::Sensor *formaldehyde_sensor_{nullptr};
